Move plugins routes into plugins folder
