+++ lib/manage-prototype-routes.js
@@ -4,15 +4,15 @@ const querystring = require('querystring')
 const csrf = require('csurf')
 
 const contextPath = '/manage-prototype'
-const router = require('../../index').requests.setupRouter(contextPath)
-const redirectingRouter = require('../../index').requests.setupRouter('/manage-prototype')
+const router = require('../index').requests.setupRouter(contextPath)
+const redirectingRouter = require('../index').requests.setupRouter('/manage-prototype')
 
-const { projectDir, packageDir } = require('../utils/paths')
+const { projectDir, packageDir } = require('./utils/paths')
 const path = require('path')
-const plugins = require('../plugins/plugins')
-const config = require('../config').getConfig()
-const { requestHttpsJson, prototypeAppScripts } = require('../utils')
-const { exec } = require('../exec')
+const plugins = require('./plugins/plugins')
+const config = require('./config').getConfig()
+const { requestHttpsJson, prototypeAppScripts } = require('./utils')
+const { exec } = require('./exec')
 
 const csrfProtection = csrf({ cookie: false })
 
@@ -25,7 +25,7 @@ const pkgPath = path.join(projectDir, 'package.json')
 
 const currentKitVersion = require(path.join(packageDir, 'package.json')).version
 const knownPlugins = require(path.join(packageDir, 'known-plugins.json'))
-const { preparePackageNameForDisplay } = require('../plugins/plugins')
+const { preparePackageNameForDisplay } = require('./plugins/plugins')
 
 const latestReleaseVersions = knownPlugins.plugins.available
   .reduce((releaseVersions, nextPlugin) => {
@@ -67,7 +67,7 @@ redirectingRouter.use((req, res, next) => {
 })
 
 // Local dependencies
-const encryptPassword = require('../utils').encryptPassword
+const encryptPassword = require('./utils').encryptPassword
 
 const password = process.env.PASSWORD
