+++ lib/plugins/plugins-routes.js
@@ -1,5 +1,5 @@
 const requests = require('../../index').requests
-const plugins = require('../plugins/plugins')
+const plugins = require('./plugins')
 
 // Serve assets from plugins
 function setupPathsFor (item) {
